KubernetesClient: Remove MobileDevice, WebDriverSession-specific objects
Use the NamespacedKubernetesClient{T} class instead.
